refactor: move `ui` folder to `src/`
