+++ src/app/[locale]/author/[id]/_components/AuthorHeader/index.tsx
@@ -1,7 +1,7 @@
 import { Body, Heading } from "@secondthunder/modsen-client-blog-ui";
 import Image from "next/image";
 
-import { Socials } from "@/components/ui";
+import { Socials } from "@/ui";
 
 import styles from "./AuthorHeader.module.scss";
 import { type AuthorHeaderProps } from "./interfaces";
