+++ src/app/[locale]/blog/[id]/_components/SimilarPosts/PostCard/index.tsx
@@ -3,10 +3,10 @@ import Image from "next/image";
 import Link from "next/link";
 import readNextPlaceholderImage from "public/assets/blog/post/read-next-placeholder.jpeg";
 
-import { AuthorLink } from "@/components/ui";
 import { routes } from "@/constants/routes";
 import { getFormattedDate } from "@/helpers/getFormattedDate";
 import { findAuthorByID } from "@/lib/authors/findAuthorByID";
+import { AuthorLink } from "@/ui";
 
 import { type PostCardProps } from "./interfaces";
 import styles from "./PostCard.module.scss";
