+++ src/ui/Avatar/Avatar.module.scss
