+++ src/ui/Avatar/Avatar.test.tsx
