+++ src/ui/Avatar/interfaces.ts
