+++ src/ui/Socials/interfaces.ts
